pyproject.toml
[project]
name = "blackjack-core"
version = "0.1.0"
description = ""
authors = [
    {name = "KutologoJM",email = "kutologomashego@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "notebook (>=7.4.3,<8.0.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
